DFTDP # 278 Move get-assistance to shared library - Moved get assistance to the shared library - Added call status, callback and preferred time models to shared api - Updated get assistance exports in api.ts file - Added callback enums in api.model file - Updated get assistance page in partner portal to use the shared component
