+++ shared-portal-ui/projects/core-ui/src/lib/api/models/call-status.ts
@@ -1,0 +1,6 @@
+/* tslint:disable */
+/* eslint-disable */
+export enum CallStatus {
+  $0 = 0,
+  $1 = 1
+}
