+++ shared-portal-ui/projects/core-ui/src/lib/api/models/index.ts
@@ -1,3 +1,6 @@
 export * from "./document";
 export * from "./driver";
 export * from "./document-sub-type"
+export * from "./call-status"
+export * from "./callback"
+export * from "./preferred-time"
