+++ shared-portal-ui/projects/core-ui/src/lib/api/models/preferred-time.ts
@@ -1,0 +1,7 @@
+/* tslint:disable */
+/* eslint-disable */
+export enum PreferredTime {
+  $0 = 0,
+  $1 = 1,
+  $2 = 2
+}
