+++ shared-portal-ui/projects/core-ui/src/lib/app.model.ts
@@ -19,3 +19,8 @@ export enum PortalsEnum{
   'MedicalPortal' = 'MedicalPortal',
   'PartnerPortal' = 'PartnerPortal'
 }
+
+export const CallStatusDescription = new Map<number | undefined, string>([
+  [0, 'Open'],
+  [1, 'Closed']
+])
