+++ shared-portal-ui/projects/core-ui/src/lib/get-assistance/cancel-callback-dialog/cancel-callback-dialog.component.html
@@ -1,0 +1,44 @@
+<div mat-dialog-title class="title">
+  <h2>Cancel Request Call Back</h2>
+  <button mat-icon-button aria-label="Close" mat-dialog-close>
+    <mat-icon>close</mat-icon>
+  </button>
+</div>
+
+<div class="dialog-content">
+  <p>Are you sure you want to cancel your Request Call Back?</p>
+  <div>
+    <mat-card>
+      <mat-card-content class="callbackData">
+        <div style="grid-area: A">
+          <label>Request Callback Date</label>
+          <div>
+            <b>April 28, 2023</b>
+          </div>
+        </div>
+
+        <div style="grid-area: B">
+          <label>Topic</label>
+          <div>
+            <b>Upload Extra Document</b>
+          </div>
+        </div>
+        <div style="grid-area: C">
+          <label>Status</label>
+          <div>
+            <b>Open</b>
+          </div>
+        </div>
+      </mat-card-content>
+    </mat-card>
+  </div>
+</div>
+
+<div mat-dialog-actions class="dialog-actions">
+  <button mat-raised-button color="primary" (click)="cancelRequestCallback()">
+    Yes, Cancel Request
+  </button>
+  <button mat-button mat-dialog-close color="primary" (click)="onCancel()">
+    No, not right now
+  </button>
+</div>
