+++ shared-portal-ui/projects/core-ui/src/lib/get-assistance/cancel-callback-dialog/cancel-callback-dialog.component.scss
@@ -1,0 +1,32 @@
+.title {
+  display: flex;
+  justify-content: space-between;
+  padding: 20px;
+  border-bottom: solid 1px gray;
+}
+
+.form-label {
+  font-family: "BCSans-Bold", "BC Sans Bold", "BC Sans", sans-serif;
+  font-weight: 700;
+  font-style: normal;
+  font-size: 16px;
+}
+
+::ng-deep .mat-dialog-container {
+  padding: 0;
+  overflow: hidden !important;
+}
+
+.dialog-content {
+  padding: 32px;
+}
+
+.callbackData {
+  display: flex;
+  justify-content: space-between;
+}
+
+.dialog-actions {
+  display: flex;
+  flex-direction: column;
+}
