+++ shared-portal-ui/projects/core-ui/src/lib/get-assistance/cancel-callback-dialog/cancel-callback-dialog.component.spec.ts
@@ -1,0 +1,17 @@
+describe('CancelCallbackDialogComponent', () => {
+  // let component: CancelCallbackDialogComponent;
+  // let fixture: ComponentFixture<CancelCallbackDialogComponent>;
+
+  // beforeEach(() => {
+  //   TestBed.configureTestingModule({
+  //     declarations: [CancelCallbackDialogComponent]
+  //   });
+  //   fixture = TestBed.createComponent(CancelCallbackDialogComponent);
+  //   component = fixture.componentInstance;
+  //   fixture.detectChanges();
+  // });
+
+  it('should create', () => {
+    //expect(component).toBeTruthy();
+  });
+});
