+++ shared-portal-ui/projects/core-ui/src/lib/get-assistance/get-assistance.component.spec.ts
@@ -1,0 +1,17 @@
+describe('GetAssistanceComponent', () => {
+  // let component: GetAssistanceComponent;
+  // let fixture: ComponentFixture<GetAssistanceComponent>;
+
+  // beforeEach(() => {
+  //   TestBed.configureTestingModule({
+  //     declarations: [GetAssistanceComponent]
+  //   });
+  //   fixture = TestBed.createComponent(GetAssistanceComponent);
+  //   component = fixture.componentInstance;
+  //   fixture.detectChanges();
+  // });
+
+  it('should create', () => {
+    //expect(component).toBeTruthy();
+  });
+});
